I think I am done with theory now. But I should read over it before I give it to patrick.
I also made up a structure for results.

also wrote the setup part, where I added a section about the quater-wave plate and the walloston prism.
I should try to find a make a better image for the wallosston prism.
I also still need a scheme of the setup.

Now I start writing on the execution part.

After that the results is on the line